refactor(check): extract expressions.ts from checker.ts
This is the largest extraction in the checker refactoring series.
expressions.ts is the HUB module - Layer 2 (Core) in the architecture.

Extracted functions (~50+):
- Main dispatch: checkExpression, checkExpressionInferred
- Literal checking: checkIntLiteral, checkFloatLiteral, checkVarRef
- Unary expressions: checkUnaryExpr, checkBitwiseNot, checkParenExpr
- Binary expressions: checkBinaryExpr, getBinaryOperands, emitLogicalOp
- Compare chains: checkCompareChain, collectCompareChainOperands
- Field access: checkFieldAccess, buildFlattenedBasePath, emitFieldAccessInst
- Index access: checkIndexAccess, validateListIndexBounds
- List literals: checkListLiteral, checkListElements, collectListElementIds

CRITICAL: checkExpression is exported and imported by checker.ts.
It only dispatches to expression-related functions, NOT to records,
match, or bindings - this prevents cyclic imports.

Also includes minor formatting fixes from biome in state.ts,
type-resolution.ts, and utils.ts.
